refactor(connect): extract wire DTOs + sentinel errors (GO-2 S3)
Move all JSON wire DTOs from connect.go / sse.go / sync.go into a
single types.go — registration, heartbeat, discovery, gateway config,
policy + route sync acks, SSE events. Struct tags unchanged; this is
a pure relocation.

Move SyncStep + newSyncStep into steps.go. Move ErrGatewayNotFound
into errors.go and add ErrNotRegistered sentinel. Replace four
anonymous fmt.Errorf("not registered") sites with the sentinel so
callers can match via errors.Is.

TestReportRouteSyncAckNotRegistered now asserts errors.Is(err,
ErrNotRegistered) instead of string-equal on the message.

LOC budget progression:
  connect.go 432 -> 383
  sync.go    456 -> 387
  sse.go     262 -> 247

go test ./internal/connect/... -race -count=1: 78 tests green.
